gdrive storage service binding

+++ src/services/storage.ts
@@ -6,13 +6,22 @@ interface WatchListData {
     driveMapping: { [key: string]: string }; // Index signature for driveMapping
 }
 
+const delay = (ms: any) => new Promise(res => setTimeout(res, ms));
+
 class WatchListService {
-    private watchListData: WatchListData;
+    private watchListData: WatchListData = {
+        toWatch: [],
+        watched: [],
+        driveMapping: {}
+    };
 
     constructor() {
         console.log('watchlist service constructor');
         this.watchListData = this.loadWatchListData();
-        this.getDriveFiles();
+    }
+
+    public async postAuth() {
+        this.syncDriveFiles();
     }
 
     private loadWatchListData(): WatchListData {
@@ -20,26 +29,70 @@ class WatchListService {
         return savedData ? JSON.parse(savedData) : { toWatch: [], watched: [], driveMapping: {} };
     }
 
-    private saveWatchListData() {
+    private save() {
         localStorage.setItem('watchListData', JSON.stringify(this.watchListData));
     }
 
     private setDriveFileId(imdbID: string, driveId: string) {
         this.watchListData.driveMapping[imdbID] = driveId;
-        this.saveWatchListData();
+        this.save();
     }
 
     private getDriveFileId(imdbID: string) {
         return this.watchListData.driveMapping[imdbID];
     }
 
-    private getDriveFiles() {
-        return gDriveService.listFilesDir().then(
-            data => {
-                if (!data) return;
-                console.log(data);
+    private _addToWatchList(id: string) {
+        if (this.watchListData.watched.includes(id))
+            this.removeWatchedList(id);
+        if (!this.watchListData.toWatch.includes(id)) {
+            this.watchListData.toWatch.push(id);
+            this.save();
+        }
+    }
+
+    private _addWatchedList(id: string) {
+        if (this.watchListData.toWatch.includes(id))
+            this.removeToWatchList(id);
+        if (!this.watchListData.watched.includes(id)) {
+            this.watchListData.watched.push(id);
+            this.save();
+        }
+    }
+
+    public async syncDriveFiles() {
+        // current behaviour delete local, drive -> local
+        // TODO conflict resolution by latest timestamp
+
+        let newToWatchList: string[] = [];
+        let newWatchedList: string[] = [];
+        let newDriveMapping: { [key: string]: string } = {};
+
+        const driveFiles = await gDriveService.listFilesDir();
+        if (!driveFiles) return;
+        driveFiles.forEach((item) => {
+            const { id, name } = item;
+            const startIndex = name.indexOf('_') + 1;
+            const endIndex = name.indexOf('.json');
+            const key = name.substring(startIndex, endIndex);
+
+            if (name.startsWith('towatch-list_')) {
+                newToWatchList.push(key);
+                newDriveMapping[key] = id;
+            } else if (name.startsWith('watched-list_')) {
+                newWatchedList.push(key);
+                newDriveMapping[key] = id;
             }
-        )
+        });
+
+        const newData: WatchListData = {
+            toWatch: newToWatchList,
+            watched: newWatchedList,
+            driveMapping: newDriveMapping
+        }
+
+        this.watchListData = newData;
+        this.save()
     }
 
     public addToWatchList(id: string) {
@@ -47,10 +100,11 @@ class WatchListService {
             this.removeWatchedList(id);
         if (!this.watchListData.toWatch.includes(id)) {
             this.watchListData.toWatch.push(id);
-            this.saveWatchListData();
-            gDriveService.addToWatchList(id).then((driveFileId: string) => {
-                this.setDriveFileId(id, driveFileId);
-            });
+            this.save();
+            if (gDriveService.isLoggedIn)
+                gDriveService.addToWatchList(id).then((driveFileId: string) => {
+                    this.setDriveFileId(id, driveFileId);
+                });
         }
     }
 
@@ -59,23 +113,26 @@ class WatchListService {
             this.removeToWatchList(id);
         if (!this.watchListData.watched.includes(id)) {
             this.watchListData.watched.push(id);
-            this.saveWatchListData();
-            gDriveService.addWatchedList(id).then((driveFileId: string) => {
-                this.setDriveFileId(id, driveFileId);
-            });
+            this.save();
+            if (gDriveService.isLoggedIn)
+                gDriveService.addWatchedList(id).then((driveFileId: string) => {
+                    this.setDriveFileId(id, driveFileId);
+                });
         }
     }
 
     public removeToWatchList(id: string) {
         this.watchListData.toWatch = this.watchListData.toWatch.filter((item) => item !== id);
-        this.saveWatchListData();
-        gDriveService.deleteFile(this.getDriveFileId(id));
+        this.save();
+        if (gDriveService.isLoggedIn)
+            gDriveService.deleteFile(this.getDriveFileId(id));
     }
 
     public removeWatchedList(id: string) {
         this.watchListData.watched = this.watchListData.watched.filter((item) => item !== id);
-        this.saveWatchListData();
-        gDriveService.deleteFile(this.getDriveFileId(id));
+        this.save();
+        if (gDriveService.isLoggedIn)
+            gDriveService.deleteFile(this.getDriveFileId(id));
     }
 
     public getWatchLists(): WatchListData {
@@ -90,7 +147,5 @@ class WatchListService {
 }
 
 const watchListService = new WatchListService();
-
-export {
-    watchListService,
-};
+export default watchListService;
+export type { WatchListData };
